Fixed some issues found during testing. Separated game_api and player_api views into different files.
